# shared/openapi/server-api.yaml
openapi: 3.1.0
info:
  title: Power Monitor Server API
  description: Central server for pm-protocol/1.0.0 sensor fleets
  version: 1.0.0
paths:
...
  /api/v1/rate-plans:
    get:
      tags:
      - rates and billing
      summary: List Rate Plans
      operationId: list_rate_plans_api_v1_rate_plans_get
      parameters:
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Rate Plans Api V1 Rate Plans Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/rate-plans/{rate_plan_id}/export:
    get:
      tags:
      - rates and billing
      summary: Export Rate Plan
      operationId: export_rate_plan_api_v1_rate_plans__rate_plan_id__export_get
      parameters:
      - name: rate_plan_id
        in: path
        required: true
        schema:
          type: string
          title: Rate Plan Id
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Export Rate Plan Api V1 Rate Plans  Rate Plan Id  Export
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/rate-plans/{rate_plan_id}/versions:
    post:
      tags:
      - rates and billing
      summary: Clone Rate Version
      operationId: clone_rate_version_api_v1_rate_plans__rate_plan_id__versions_post
      parameters:
      - name: rate_plan_id
        in: path
        required: true
        schema:
          type: string
          title: Rate Plan Id
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Payload
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Clone Rate Version Api V1 Rate Plans  Rate Plan Id  Versions
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/rate-versions/{version_id}/activate:
    post:
      tags:
      - rates and billing
      summary: Activate Rate Version
      operationId: activate_rate_version_api_v1_rate_versions__version_id__activate_post
      parameters:
      - name: version_id
        in: path
        required: true
        schema:
          type: string
          title: Version Id
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: boolean
                title: Response Activate Rate Version Api V1 Rate Versions  Version
                  Id  Activate Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/rates/preview:
    post:
      tags:
      - rates and billing
      summary: Preview Rate
      operationId: preview_rate_api_v1_rates_preview_post
      parameters:
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatePreviewRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RatePreviewResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/billing/cycles:
    get:
      tags:
      - rates and billing
      summary: Billing Cycles
      operationId: billing_cycles_api_v1_billing_cycles_get
      parameters:
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response Billing Cycles Api V1 Billing Cycles Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/billing/runs:
    get:
      tags:
      - rates and billing
      summary: Billing Runs
      operationId: billing_runs_api_v1_billing_runs_get
      parameters:
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response Billing Runs Api V1 Billing Runs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/billing/recalculations:
    post:
      tags:
      - rates and billing
      summary: Queue Recalculation
      operationId: queue_recalculation_api_v1_billing_recalculations_post
      parameters:
      - name: X-CSRF-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Csrf-Token
      - name: pm_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Pm Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostRecalculationRequest'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Queue Recalculation Api V1 Billing Recalculations
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CostComponent:
      properties:
        name:
          type: string
          title: Name
        amount:
          type: string
          title: Amount
      additionalProperties: false
      type: object
      required:
      - name
      - amount
      title: CostComponent
    CostRecalculationRequest:
      properties:
        utility_account_id:
          type: string
          title: Utility Account Id
        aggregate_set_id:
          type: string
          title: Aggregate Set Id
        rate_version_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Rate Version Id
        input_start:
          type: string
          format: date-time
          title: Input Start
        input_end:
          type: string
          format: date-time
          title: Input End
      additionalProperties: false
      type: object
      required:
      - utility_account_id
      - aggregate_set_id
      - input_start
      - input_end
      title: CostRecalculationRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RatePreviewRequest:
      properties:
        plan_code:
          type: string
          title: Plan Code
        interval_start:
          type: string
          format: date-time
          title: Interval Start
        interval_end:
          type: string
          format: date-time
          title: Interval End
        energy_kwh:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
          title: Energy Kwh
        cost_scope:
          type: string
          enum:
          - energy_only
          - allocated_account
          - full_account
          title: Cost Scope
          default: energy_only
        baseline_allocation_kwh:
          anyOf:
          - type: number
            minimum: 0.0
          - type: string
          - type: 'null'
          title: Baseline Allocation Kwh
        billing_days:
          type: integer
          maximum: 366.0
          minimum: 1.0
          title: Billing Days
          default: 1
        cca_adjustment_per_kwh:
          anyOf:
          - type: number
          - type: string
          title: Cca Adjustment Per Kwh
          default: '0'
        other_adjustment:
          anyOf:
          - type: number
          - type: string
          title: Other Adjustment
          default: '0'
      additionalProperties: false
      type: object
      required:
      - plan_code
      - interval_start
      - interval_end
      - energy_kwh
      title: RatePreviewRequest
    RatePreviewResponse:
      properties:
        plan_code:
          type: string
          title: Plan Code
        rate_version:
          type: string
          title: Rate Version
        timezone:
          type: string
          title: Timezone
        energy_by_bucket_kwh:
          additionalProperties:
            type: string
          type: object
          title: Energy By Bucket Kwh
        components:
          items:
            $ref: '#/components/schemas/CostComponent'
          type: array
          title: Components
        unrounded_total:
          type: string
          title: Unrounded Total
        display_total:
          type: string
          title: Display Total
        coverage_percent:
          type: string
          title: Coverage Percent
        disclosure:
          type: string
          title: Disclosure
      additionalProperties: false
      type: object
      required:
      - plan_code
      - rate_version
      - timezone
      - energy_by_bucket_kwh
      - components
      - unrounded_total
      - display_total
      - coverage_percent
      - disclosure
      title: RatePreviewResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError